refactored all scss files to have absolute path and calling each files appropriately
