+++ client/src/components/About/AboutMe/AboutMe.module.scss
@@ -1,4 +1,4 @@
-@import './App.scss';
+@import '../../App.module.scss';
 
 @font-face {
   font-family: 'Edmonsans-Bold';
