+++ client/src/components/About/Details/Details.module.scss
@@ -1,4 +1,4 @@
-@import './App.scss';
+@import '../../App.module.scss';
 
 .detailsBg {
   // font-size: 21px;
