+++ client/src/components/About/Technical/Resume/Resume.module.scss
@@ -1,4 +1,4 @@
-@import './App.scss';
+@import '../../../App.module.scss';
 
 body {
   background: $tim-default-bg;
