+++ client/src/components/About/Technical/styles/Technical.module.scss
@@ -1,5 +1,5 @@
-@import './App.scss';
-@import './HrProjects.scss';
+@import '../../../App.module.scss';
+@import '../../../Projects/HR/HrProjects.module.scss';
 // .headingBg {
 //   color: #fff;
 //   transition: all 0.5s;
