+++ client/src/components/Footer/Footer.module.scss
@@ -1,4 +1,4 @@
-@import './App.scss';
+@import '../App.module.scss';
 
 .iconBg {
   margin-left: 0.2rem;
