+++ client/src/components/Projects/HR/HrProjects.module.scss
@@ -1,4 +1,4 @@
-@import './App.scss';
+@import '../../App.module.scss';
 
 $modal-bg: #e3e2e0;
