+++ client/src/styles/main.scss
@@ -1,1 +1,13 @@
 @import 'custom';
+
+@import 'client/src/components/App.module.scss';
+@import 'client/src/components/Navigation/NavBar/NavBar.module.scss';
+@import 'client/src/components/About/AboutMe/AboutMe.module.scss';
+@import 'client/src/components/About/Articles/Articles.module.scss';
+@import 'client/src/components/About/Details/Details.module.scss';
+@import 'client/src/components/About/Technical/styles/Technical.module.scss';
+@import 'client/src/components/About/Technical/Resume/Resume.module.scss';
+@import 'client/src/components/Footer/Footer.module.scss';
+@import 'client/src/components/Header/Header.module.scss';
+@import 'client/src/components/Navigation/Headline/Headline.module.scss';
+@import 'client/src/components/Projects/HR/HrProjects.module.scss';
